+++ core/tests/connection_limits.rs
@@ -25,7 +25,7 @@ use libp2p_core::multiaddr::{multiaddr, Multiaddr};
 use libp2p_core::{
     connection::PendingConnectionError,
     network::{ConnectionLimits, DialError, NetworkConfig, NetworkEvent},
-    PeerId,
+    DialOpts, PeerId,
 };
 use quickcheck::*;
 use std::task::Poll;
@@ -46,15 +46,23 @@ fn max_outgoing() {
     let target = PeerId::random();
     for _ in 0..outgoing_limit {
         network
-            .peer(target.clone())
-            .dial(vec![addr.clone()], TestHandler())
+            .dial(
+                TestHandler(),
+                DialOpts::peer_id(target)
+                    .addresses(vec![addr.clone()])
+                    .build(),
+            )
             .ok()
             .expect("Unexpected connection limit.");
     }
 
     match network
-        .peer(target.clone())
-        .dial(vec![addr.clone()], TestHandler())
+        .dial(
+            TestHandler(),
+            DialOpts::peer_id(target)
+                .addresses(vec![addr.clone()])
+                .build(),
+        )
         .expect_err("Unexpected dialing success.")
     {
         DialError::ConnectionLimit { limit, handler: _ } => {
@@ -122,7 +130,7 @@ fn max_established_incoming() {
         // Spawn and block on the dialer.
         async_std::task::block_on({
             let mut n = 0;
-            let _ = network2.dial(&listen_addr, TestHandler()).unwrap();
+            let _ = network2.dial(TestHandler(), listen_addr.clone()).unwrap();
 
             let mut expected_closed = false;
             let mut network_1_established = false;
@@ -188,7 +196,7 @@ fn max_established_incoming() {
                         if n <= limit {
                             // Dial again until the limit is exceeded.
                             n += 1;
-                            network2.dial(&listen_addr, TestHandler()).unwrap();
+                            network2.dial(TestHandler(), listen_addr.clone()).unwrap();
 
                             if n == limit {
                                 // The the next dialing attempt exceeds the limit, this