+++ core/tests/network_dial_error.rs
@@ -22,11 +22,12 @@ mod util;
 
 use futures::prelude::*;
 use libp2p_core::multiaddr::multiaddr;
+use libp2p_core::DialOpts;
 use libp2p_core::{
     connection::PendingConnectionError,
     multiaddr::Protocol,
     network::{NetworkConfig, NetworkEvent},
-    PeerId,
+    ConnectedPoint, Endpoint, PeerId,
 };
 use rand::seq::SliceRandom;
 use std::{io, task::Poll};
@@ -41,7 +42,7 @@ fn deny_incoming_connec() {
 
     swarm1.listen_on("/memory/0".parse().unwrap()).unwrap();
 
-    let address = async_std::task::block_on(future::poll_fn(|cx| match swarm1.poll(cx) {
+    let address = futures::executor::block_on(future::poll_fn(|cx| match swarm1.poll(cx) {
         Poll::Ready(NetworkEvent::NewListenerAddress { listen_addr, .. }) => {
             Poll::Ready(listen_addr)
         }
@@ -50,11 +51,15 @@ fn deny_incoming_connec() {
     }));
 
     swarm2
-        .peer(swarm1.local_peer_id().clone())
-        .dial(vec![address.clone()], TestHandler())
+        .dial(
+            TestHandler(),
+            DialOpts::peer_id(*swarm1.local_peer_id())
+                .addresses(vec![address.clone()])
+                .build(),
+        )
         .unwrap();
 
-    async_std::task::block_on(future::poll_fn(|cx| -> Poll<Result<(), io::Error>> {
+    futures::executor::block_on(future::poll_fn(|cx| -> Poll<Result<(), io::Error>> {
         match swarm1.poll(cx) {
             Poll::Ready(NetworkEvent::IncomingConnection { connection, .. }) => drop(connection),
             Poll::Ready(_) => unreachable!(),
@@ -84,6 +89,58 @@ fn deny_incoming_connec() {
 }
 
 #[test]
+fn invalid_peer_id() {
+    // Checks whether dialing an address containing the wrong peer id raises an error
+    // for the expected peer id instead of the obtained peer id.
+
+    let mut swarm1 = test_network(NetworkConfig::default());
+    let mut swarm2 = test_network(NetworkConfig::default());
+
+    swarm1.listen_on("/memory/0".parse().unwrap()).unwrap();
+
+    let address = futures::executor::block_on(future::poll_fn(|cx| match swarm1.poll(cx) {
+        Poll::Ready(NetworkEvent::NewListenerAddress { listen_addr, .. }) => {
+            Poll::Ready(listen_addr)
+        }
+        Poll::Pending => Poll::Pending,
+        _ => panic!("Was expecting the listen address to be reported"),
+    }));
+
+    let other_id = PeerId::random();
+    let other_addr = address.with(Protocol::P2p(other_id.into()));
+
+    swarm2.dial(TestHandler(), other_addr.clone()).unwrap();
+
+    let (peer_id, error) = futures::executor::block_on(future::poll_fn(|cx| {
+        if let Poll::Ready(NetworkEvent::IncomingConnection { connection, .. }) = swarm1.poll(cx) {
+            swarm1.accept(connection, TestHandler()).unwrap();
+        }
+
+        match swarm2.poll(cx) {
+            Poll::Ready(NetworkEvent::DialError { peer_id, error, .. }) => {
+                Poll::Ready((peer_id, error))
+            }
+            Poll::Ready(x) => panic!("unexpected {:?}", x),
+            Poll::Pending => Poll::Pending,
+        }
+    }));
+    assert_eq!(peer_id, other_id);
+    match error {
+        PendingConnectionError::WrongPeerId { obtained, endpoint } => {
+            assert_eq!(obtained, *swarm1.local_peer_id());
+            assert_eq!(
+                endpoint,
+                ConnectedPoint::Dialer {
+                    address: other_addr,
+                    role_override: Endpoint::Dialer,
+                }
+            );
+        }
+        x => panic!("wrong error {:?}", x),
+    }
+}
+
+#[test]
 fn dial_self() {
     // Check whether dialing ourselves correctly fails.
     //
@@ -98,7 +155,7 @@ fn dial_self() {
     let mut swarm = test_network(NetworkConfig::default());
     swarm.listen_on("/memory/0".parse().unwrap()).unwrap();
 
-    let local_address = async_std::task::block_on(future::poll_fn(|cx| match swarm.poll(cx) {
+    let local_address = futures::executor::block_on(future::poll_fn(|cx| match swarm.poll(cx) {
         Poll::Ready(NetworkEvent::NewListenerAddress { listen_addr, .. }) => {
             Poll::Ready(listen_addr)
         }
@@ -106,16 +163,16 @@ fn dial_self() {
         _ => panic!("Was expecting the listen address to be reported"),
     }));
 
-    swarm.dial(&local_address, TestHandler()).unwrap();
+    swarm.dial(TestHandler(), local_address.clone()).unwrap();
 
     let mut got_dial_err = false;
     let mut got_inc_err = false;
-    async_std::task::block_on(future::poll_fn(|cx| -> Poll<Result<(), io::Error>> {
+    futures::executor::block_on(future::poll_fn(|cx| -> Poll<Result<(), io::Error>> {
         loop {
             match swarm.poll(cx) {
                 Poll::Ready(NetworkEvent::DialError {
                     peer_id,
-                    error: PendingConnectionError::InvalidPeerId { .. },
+                    error: PendingConnectionError::WrongPeerId { .. },
                     ..
                 }) => {
                     assert_eq!(&peer_id, swarm.local_peer_id());
@@ -152,7 +209,7 @@ fn dial_self_by_id() {
     // Trying to dial self by passing the same `PeerId` shouldn't even be possible in the first
     // place.
     let mut swarm = test_network(NetworkConfig::default());
-    let peer_id = swarm.local_peer_id().clone();
+    let peer_id = *swarm.local_peer_id();
     assert!(swarm.peer(peer_id).into_disconnected().is_none());
 }
 
@@ -174,11 +231,15 @@ fn multiple_addresses_err() {
     addresses.shuffle(&mut rand::thread_rng());
 
     swarm
-        .peer(target.clone())
-        .dial(addresses.clone(), TestHandler())
+        .dial(
+            TestHandler(),
+            DialOpts::peer_id(target)
+                .addresses(addresses.clone())
+                .build(),
+        )
         .unwrap();
 
-    async_std::task::block_on(future::poll_fn(|cx| -> Poll<Result<(), io::Error>> {
+    futures::executor::block_on(future::poll_fn(|cx| -> Poll<Result<(), io::Error>> {
         loop {
             match swarm.poll(cx) {
                 Poll::Ready(NetworkEvent::DialError {