+++ protocols/relay/src/v1/transport.rs
@@ -18,14 +18,14 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-use crate::behaviour::{BehaviourToListenerMsg, OutgoingRelayReqError};
-use crate::protocol;
-use crate::RequestId;
+use crate::v1::behaviour::BehaviourToListenerMsg;
+use crate::v1::{Connection, RequestId};
 use futures::channel::mpsc;
 use futures::channel::oneshot;
 use futures::future::{BoxFuture, Future, FutureExt};
 use futures::sink::SinkExt;
 use futures::stream::{Stream, StreamExt};
+use libp2p_core::connection::Endpoint;
 use libp2p_core::either::{EitherError, EitherFuture, EitherOutput};
 use libp2p_core::multiaddr::{Multiaddr, Protocol};
 use libp2p_core::transport::{ListenerEvent, TransportError};
@@ -43,7 +43,7 @@ use std::task::{Context, Poll};
 ///    ```
 ///    # use libp2p_core::{Multiaddr, multiaddr::{Protocol}, Transport};
 ///    # use libp2p_core::transport::memory::MemoryTransport;
-///    # use libp2p_relay::{RelayConfig, new_transport_and_behaviour};
+///    # use libp2p_relay::v1::{RelayConfig, new_transport_and_behaviour};
 ///    # let inner_transport = MemoryTransport::default();
 ///    # let (relay_transport, relay_behaviour) = new_transport_and_behaviour(
 ///    #     RelayConfig::default(),
@@ -57,7 +57,7 @@ use std::task::{Context, Poll};
 ///    ```
 ///    # use libp2p_core::{Multiaddr, multiaddr::{Protocol}, PeerId, Transport};
 ///    # use libp2p_core::transport::memory::MemoryTransport;
-///    # use libp2p_relay::{RelayConfig, new_transport_and_behaviour};
+///    # use libp2p_relay::v1::{RelayConfig, new_transport_and_behaviour};
 ///    # let inner_transport = MemoryTransport::default();
 ///    # let (relay_transport, relay_behaviour) = new_transport_and_behaviour(
 ///    #     RelayConfig::default(),
@@ -77,7 +77,7 @@ use std::task::{Context, Poll};
 ///    ```
 ///    # use libp2p_core::{Multiaddr, multiaddr::{Protocol}, PeerId, Transport};
 ///    # use libp2p_core::transport::memory::MemoryTransport;
-///    # use libp2p_relay::{RelayConfig, new_transport_and_behaviour};
+///    # use libp2p_relay::v1::{RelayConfig, new_transport_and_behaviour};
 ///    # let inner_transport = MemoryTransport::default();
 ///    # let (relay_transport, relay_behaviour) = new_transport_and_behaviour(
 ///    #     RelayConfig::default(),
@@ -98,7 +98,7 @@ use std::task::{Context, Poll};
 ///    ```
 ///    # use libp2p_core::{Multiaddr, multiaddr::{Protocol}, PeerId, Transport};
 ///    # use libp2p_core::transport::memory::MemoryTransport;
-///    # use libp2p_relay::{RelayConfig, new_transport_and_behaviour};
+///    # use libp2p_relay::v1::{RelayConfig, new_transport_and_behaviour};
 ///    # let inner_transport = MemoryTransport::default();
 ///    # let (relay_transport, relay_behaviour) = new_transport_and_behaviour(
 ///    #     RelayConfig::default(),
@@ -121,7 +121,7 @@ impl<T: Clone> RelayTransport<T> {
     ///
     ///```
     /// # use libp2p_core::transport::dummy::DummyTransport;
-    /// # use libp2p_relay::{RelayConfig, new_transport_and_behaviour};
+    /// # use libp2p_relay::v1::{RelayConfig, new_transport_and_behaviour};
     ///
     /// let inner_transport = DummyTransport::<()>::new();
     /// let (relay_transport, relay_behaviour) = new_transport_and_behaviour(
@@ -143,7 +143,7 @@ impl<T: Clone> RelayTransport<T> {
 }
 
 impl<T: Transport + Clone> Transport for RelayTransport<T> {
-    type Output = EitherOutput<<T as Transport>::Output, protocol::Connection>;
+    type Output = EitherOutput<<T as Transport>::Output, Connection>;
     type Error = EitherError<<T as Transport>::Error, RelayError>;
     type Listener = RelayListener<T>;
     type ListenerUpgrade = RelayedListenerUpgrade<T>;
@@ -221,15 +221,41 @@ impl<T: Transport + Clone> Transport for RelayTransport<T> {
     }
 
     fn dial(self, addr: Multiaddr) -> Result<Self::Dial, TransportError<Self::Error>> {
+        self.do_dial(addr, Endpoint::Dialer)
+    }
+
+    fn dial_as_listener(self, addr: Multiaddr) -> Result<Self::Dial, TransportError<Self::Error>> {
+        self.do_dial(addr, Endpoint::Listener)
+    }
+
+    fn address_translation(&self, server: &Multiaddr, observed: &Multiaddr) -> Option<Multiaddr> {
+        self.inner_transport.address_translation(server, observed)
+    }
+}
+
+impl<T: Transport + Clone> RelayTransport<T> {
+    fn do_dial(
+        self,
+        addr: Multiaddr,
+        role_override: Endpoint,
+    ) -> Result<<Self as Transport>::Dial, TransportError<<Self as Transport>::Error>> {
         match parse_relayed_multiaddr(addr)? {
             // Address does not contain circuit relay protocol. Use inner transport.
-            Err(addr) => match self.inner_transport.dial(addr) {
-                Ok(dialer) => Ok(EitherFuture::First(dialer)),
-                Err(TransportError::MultiaddrNotSupported(addr)) => {
-                    Err(TransportError::MultiaddrNotSupported(addr))
+            Err(addr) => {
+                let dial = match role_override {
+                    Endpoint::Dialer => self.inner_transport.dial(addr),
+                    Endpoint::Listener => self.inner_transport.dial_as_listener(addr),
+                };
+                match dial {
+                    Ok(dialer) => Ok(EitherFuture::First(dialer)),
+                    Err(TransportError::MultiaddrNotSupported(addr)) => {
+                        Err(TransportError::MultiaddrNotSupported(addr))
+                    }
+                    Err(TransportError::Other(err)) => {
+                        Err(TransportError::Other(EitherError::A(err)))
+                    }
                 }
-                Err(TransportError::Other(err)) => Err(TransportError::Other(EitherError::A(err))),
-            },
+            }
             // Address does contain circuit relay protocol. Dial destination via relay.
             Ok(RelayedMultiaddr {
                 relay_peer_id,
@@ -263,10 +289,6 @@ impl<T: Transport + Clone> Transport for RelayTransport<T> {
                 ))
             }
         }
-    }
-
-    fn address_translation(&self, server: &Multiaddr, observed: &Multiaddr) -> Option<Multiaddr> {
-        self.inner_transport.address_translation(server, observed)
     }
 }
 
@@ -427,17 +449,17 @@ impl<T: Transport> Stream for RelayListener<T> {
     }
 }
 
-pub type RelayedDial = BoxFuture<'static, Result<protocol::Connection, RelayError>>;
+pub type RelayedDial = BoxFuture<'static, Result<Connection, RelayError>>;
 
 #[pin_project(project = RelayedListenerUpgradeProj)]
 pub enum RelayedListenerUpgrade<T: Transport> {
     Inner(#[pin] <T as Transport>::ListenerUpgrade),
-    Relayed(Option<protocol::Connection>),
+    Relayed(Option<Connection>),
 }
 
 impl<T: Transport> Future for RelayedListenerUpgrade<T> {
     type Output = Result<
-        EitherOutput<<T as Transport>::Output, protocol::Connection>,
+        EitherOutput<<T as Transport>::Output, Connection>,
         EitherError<<T as Transport>::Error, RelayError>,
     >;
     fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
@@ -534,7 +556,7 @@ impl std::fmt::Display for RelayError {
 
 impl std::error::Error for RelayError {}
 
-/// Message from the [`RelayTransport`] to the [`Relay`](crate::Relay)
+/// Message from the [`RelayTransport`] to the [`Relay`](crate::v1::Relay)
 /// [`NetworkBehaviour`](libp2p_swarm::NetworkBehaviour).
 pub enum TransportToBehaviourMsg {
     /// Dial destination node via relay node.
@@ -544,7 +566,7 @@ pub enum TransportToBehaviourMsg {
         relay_peer_id: PeerId,
         dst_addr: Option<Multiaddr>,
         dst_peer_id: PeerId,
-        send_back: oneshot::Sender<Result<protocol::Connection, OutgoingRelayReqError>>,
+        send_back: oneshot::Sender<Result<Connection, OutgoingRelayReqError>>,
     },
     /// Listen for incoming relayed connections via relay node.
     ListenReq {
@@ -554,4 +576,9 @@ pub enum TransportToBehaviourMsg {
         relay_peer_id_and_addr: Option<(PeerId, Multiaddr)>,
         to_listener: mpsc::Sender<BehaviourToListenerMsg>,
     },
+}
+
+#[derive(Debug, Eq, PartialEq)]
+pub enum OutgoingRelayReqError {
+    DialingRelay,
 }